Add unit tests Add XML comments
