+++ msal/tests/Test.MSAL.NET.Unit/PublicClientApplicationTests.cs
@@ -25,27 +25,28 @@
 //
 //------------------------------------------------------------------------------
 
+using Microsoft.Identity.Client;
+using Microsoft.Identity.Client.Internal;
+using Microsoft.Identity.Core;
+using Microsoft.Identity.Core.Cache;
+using Microsoft.Identity.Core.Helpers;
+using Microsoft.Identity.Core.Http;
+using Microsoft.Identity.Core.Instance;
+using Microsoft.Identity.Core.OAuth2;
+using Microsoft.Identity.Core.Telemetry;
+using Microsoft.Identity.Core.UI;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+using NSubstitute;
 using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Net;
 using System.Net.Http;
-using System.Threading.Tasks;
-using Microsoft.Identity.Client;
-using Microsoft.Identity.Client.Internal;
-using Microsoft.Identity.Core.Cache;
-using Microsoft.Identity.Core.Helpers;
-using Microsoft.Identity.Core.Http;
-using Microsoft.Identity.Core.Instance;
-using Microsoft.Identity.Core.OAuth2;
-using Microsoft.Identity.Core.UI;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
-using NSubstitute;
-using Microsoft.Identity.Core.Telemetry;
-using Test.MSAL.NET.Unit.Mocks;
-using Test.Microsoft.Identity.Core.Unit.Mocks;
 using System.Security;
+using System.Threading.Tasks;
+using Test.Microsoft.Identity.Core.Unit.Mocks;
+using Test.MSAL.NET.Unit.Mocks;
 
 namespace Test.MSAL.NET.Unit
 {
@@ -333,360 +334,6 @@ namespace Test.MSAL.NET.Unit
             Assert.AreEqual(TestConstants.UniqueId, result.UniqueId);
             Assert.AreEqual(TestConstants.CreateUserIdentifer(), result.Account.HomeAccountId);
             Assert.AreEqual(TestConstants.DisplayableId, result.Account.Username);
-
-            Assert.IsTrue(HttpMessageHandlerFactory.IsMocksQueueEmpty, "All mocks should have been consumed");
-        }
-
-        [TestMethod]
-        [TestCategory("PublicClientApplicationTests")]
-        [DeploymentItem(@"Resources\TestMex.xml", "MsalResource")]
-        [DeploymentItem(@"Resources\WsTrustResponse13.xml", "MsalResource")]
-        public async Task AcquireTokenByIntegratedWindowsAuthTest()
-        {
-            MockWebUI ui = new MockWebUI()
-            {
-                MockResult = new AuthorizationResult(AuthorizationStatus.Success,
-                  TestConstants.AuthorityHomeTenant + "?code=some-code")
-            };
-
-            //add mock response for tenant endpoint discovery
-            HttpMessageHandlerFactory.AddMockHandler(new MockHttpMessageHandler
-            {
-                Method = HttpMethod.Get,
-                ResponseMessage = MockHelpers.CreateOpenIdConfigurationResponse(TestConstants.AuthorityHomeTenant)
-            });
-
-            // user realm discovery
-            HttpMessageHandlerFactory.AddMockHandler(new MockHttpMessageHandler
-            {
-                Method = HttpMethod.Get,
-                ResponseMessage = new HttpResponseMessage(HttpStatusCode.OK)
-                {
-                    Content = new StringContent("{\"ver\":\"1.0\",\"account_type\":\"federated\",\"domain_name\":\"microsoft.com\"," +
-                                                "\"federation_protocol\":\"WSTrust\",\"federation_metadata_url\":" +
-                                                "\"https://msft.sts.microsoft.com/adfs/services/trust/mex\"," +
-                                                "\"federation_active_auth_url\":\"https://msft.sts.microsoft.com/adfs/services/trust/2005/usernamemixed\"" +
-                                                ",\"cloud_instance_name\":\"login.microsoftonline.com\"}")
-                }
-            });
-
-            // MEX
-            HttpMessageHandlerFactory.AddMockHandler(new MockHttpMessageHandler
-            {
-                Url = "https://msft.sts.microsoft.com/adfs/services/trust/mex",
-                Method = HttpMethod.Get,
-                ResponseMessage = new HttpResponseMessage(HttpStatusCode.OK)
-                {
-                    Content = new StringContent(File.ReadAllText(@"MsalResource\TestMex.xml"))
-                }
-            });
-
-            // WsTrust
-            HttpMessageHandlerFactory.AddMockHandler(new MockHttpMessageHandler
-            {
-                Url = "https://msft.sts.microsoft.com/adfs/services/trust/13/windowstransport",
-                Method = HttpMethod.Post,
-                ResponseMessage = new HttpResponseMessage(HttpStatusCode.OK)
-                {
-                    Content = new StringContent(File.ReadAllText(@"MsalResource\WsTrustResponse13.xml"))
-                }
-            });
-
-            // AAD
-            HttpMessageHandlerFactory.AddMockHandler(new MockHttpMessageHandler
-            {
-                Url = "https://login.microsoftonline.com/home/oauth2/v2.0/token",
-                Method = HttpMethod.Post,
-                PostData = new Dictionary<string, string>()
-                {
-                    {"grant_type", "urn:ietf:params:oauth:grant-type:saml1_1-bearer"},
-                    {"scope", "offline_access openid profile r1/scope1 r1/scope2"}
-                },
-                ResponseMessage = MockHelpers.CreateSuccessTokenResponseMessage()
-            });
-
-            PublicClientApplication app = new PublicClientApplication(TestConstants.ClientId);
-            AuthenticationResult result = await app.AcquireTokenByIntegratedWindowsAuthAsync(TestConstants.Scope, TestConstants.User.Username).ConfigureAwait(false);
-
-            Assert.IsNotNull(result);
-            Assert.AreEqual("some-access-token", result.AccessToken);
-            Assert.IsNotNull(result.Account);
-            Assert.AreEqual(TestConstants.DisplayableId, result.Account.Username);
-
-            Assert.IsTrue(HttpMessageHandlerFactory.IsMocksQueueEmpty, "All mocks should have been consumed");
-        }
-
-        [TestMethod]
-        [TestCategory("PublicClientApplicationTests")]
-        [DeploymentItem(@"Resources\TestMex.xml", "MsalResource")]
-        [DeploymentItem(@"Resources\WsTrustResponse.xml", "MsalResource")]
-        public async Task FederatedUsernamePasswordAcquireTokenTest()
-        {
-            MockWebUI ui = new MockWebUI()
-            {
-                MockResult = new AuthorizationResult(AuthorizationStatus.Success,
-                 TestConstants.AuthorityOrganizationsTenant + "?code=some-code")
-            };
-
-            //add mock response for tenant endpoint discovery
-            HttpMessageHandlerFactory.AddMockHandler(new MockHttpMessageHandler
-            {
-                Method = HttpMethod.Get,
-                ResponseMessage = MockHelpers.CreateOpenIdConfigurationResponse(TestConstants.AuthorityOrganizationsTenant)
-            });
-
-            // user realm discovery
-            HttpMessageHandlerFactory.AddMockHandler(new MockHttpMessageHandler
-            {
-                Method = HttpMethod.Get,
-                ResponseMessage = new HttpResponseMessage(HttpStatusCode.OK)
-                {
-                    Content = new StringContent("{\"ver\":\"1.0\",\"account_type\":\"federated\",\"domain_name\":\"microsoft.com\"," +
-                                                "\"federation_protocol\":\"WSTrust\",\"federation_metadata_url\":" +
-                                                "\"https://msft.sts.microsoft.com/adfs/services/trust/mex\"," +
-                                                "\"federation_active_auth_url\":\"https://msft.sts.microsoft.com/adfs/services/trust/2005/usernamemixed\"" +
-                                                ",\"cloud_instance_name\":\"login.microsoftonline.com\"}")
-                }
-            });
-
-            // MEX
-            HttpMessageHandlerFactory.AddMockHandler(new MockHttpMessageHandler
-            {
-                Url = "https://msft.sts.microsoft.com/adfs/services/trust/mex",
-                Method = HttpMethod.Get,
-                ResponseMessage = new HttpResponseMessage(HttpStatusCode.OK)
-                {
-                    Content = new StringContent(File.ReadAllText(@"MsalResource\TestMex.xml"))
-                }
-            });
-
-            // WsTrust
-            HttpMessageHandlerFactory.AddMockHandler(new MockHttpMessageHandler
-            {
-                Url = "https://msft.sts.microsoft.com/adfs/services/trust/2005/usernamemixed",
-                Method = HttpMethod.Post,
-                ResponseMessage = new HttpResponseMessage(HttpStatusCode.OK)
-                {
-                    Content = new StringContent(File.ReadAllText(@"MsalResource\WsTrustResponse.xml"))
-                }
-            });
-
-            // AAD
-            HttpMessageHandlerFactory.AddMockHandler(new MockHttpMessageHandler
-            {
-                Url = "https://login.microsoftonline.com/organizations/oauth2/v2.0/token",
-                Method = HttpMethod.Post,
-                PostData = new Dictionary<string, string>()
-                {
-                    {"grant_type", "urn:ietf:params:oauth:grant-type:saml1_1-bearer"},
-                    {"scope", "offline_access openid profile r1/scope1 r1/scope2"}
-                },
-                ResponseMessage = MockHelpers.CreateSuccessTokenResponseMessage()
-            });
-
-            PublicClientApplication app = new PublicClientApplication(TestConstants.ClientId);
-
-            AuthenticationResult result = await app.AcquireTokenByUsernamePasswordAsync(TestConstants.Scope, TestConstants.User.Username, TestConstants.DefaultPassword).ConfigureAwait(false);
-
-            Assert.IsNotNull(result);
-            Assert.AreEqual("some-access-token", result.AccessToken);
-            Assert.IsNotNull(result.Account);
-            Assert.AreEqual(TestConstants.User.Username, result.Account.Username);
-
-            Assert.IsTrue(HttpMessageHandlerFactory.IsMocksQueueEmpty, "All mocks should have been consumed");
-        }
-
-        [TestMethod]
-        [TestCategory("PublicClientApplicationTests")]
-        [DeploymentItem(@"Resources\TestMex.xml", "MsalResource")]
-        [DeploymentItem(@"Resources\WsTrustResponse.xml", "MsalResource")]
-        public async Task FederatedUsernamePasswordWithSecureStringAcquireTokenTest()
-        {
-            MockWebUI ui = new MockWebUI()
-            {
-                MockResult = new AuthorizationResult(AuthorizationStatus.Success,
-                 TestConstants.AuthorityOrganizationsTenant + "?code=some-code")
-            };
-
-            //add mock response for tenant endpoint discovery
-            HttpMessageHandlerFactory.AddMockHandler(new MockHttpMessageHandler
-            {
-                Method = HttpMethod.Get,
-                ResponseMessage = MockHelpers.CreateOpenIdConfigurationResponse(TestConstants.AuthorityOrganizationsTenant)
-            });
-
-            // user realm discovery
-            HttpMessageHandlerFactory.AddMockHandler(new MockHttpMessageHandler
-            {
-                Method = HttpMethod.Get,
-                ResponseMessage = new HttpResponseMessage(HttpStatusCode.OK)
-                {
-                    Content = new StringContent("{\"ver\":\"1.0\",\"account_type\":\"federated\",\"domain_name\":\"microsoft.com\"," +
-                                                "\"federation_protocol\":\"WSTrust\",\"federation_metadata_url\":" +
-                                                "\"https://msft.sts.microsoft.com/adfs/services/trust/mex\"," +
-                                                "\"federation_active_auth_url\":\"https://msft.sts.microsoft.com/adfs/services/trust/2005/usernamemixed\"" +
-                                                ",\"cloud_instance_name\":\"login.microsoftonline.com\"}")
-                },
-                QueryParams = new Dictionary<string, string>()
-                {
-                    {"api-version", "1.0"}
-                }
-            });
-
-            // MEX
-            HttpMessageHandlerFactory.AddMockHandler(new MockHttpMessageHandler
-            {
-                Url = "https://msft.sts.microsoft.com/adfs/services/trust/mex",
-                Method = HttpMethod.Get,
-                ResponseMessage = new HttpResponseMessage(HttpStatusCode.OK)
-                {
-                    Content = new StringContent(File.ReadAllText(@"MsalResource\TestMex.xml"))
-                }
-            });
-
-            // WsTrust
-            HttpMessageHandlerFactory.AddMockHandler(new MockHttpMessageHandler
-            {
-                Url = "https://msft.sts.microsoft.com/adfs/services/trust/2005/usernamemixed",
-                Method = HttpMethod.Post,
-                ResponseMessage = new HttpResponseMessage(HttpStatusCode.OK)
-                {
-                    Content = new StringContent(File.ReadAllText(@"MsalResource\WsTrustResponse.xml"))
-                }
-            });
-
-            // AAD
-            HttpMessageHandlerFactory.AddMockHandler(new MockHttpMessageHandler
-            {
-                Url = "https://login.microsoftonline.com/organizations/oauth2/v2.0/token",
-                Method = HttpMethod.Post,
-                PostData = new Dictionary<string, string>()
-                {
-                    {"grant_type", "urn:ietf:params:oauth:grant-type:saml1_1-bearer"},
-                    {"scope", "offline_access openid profile r1/scope1 r1/scope2"}
-                },
-                ResponseMessage = MockHelpers.CreateSuccessTokenResponseMessage()
-            });
-
-            PublicClientApplication app = new PublicClientApplication(TestConstants.ClientId);
-
-            SecureString str = new SecureString();
-            str.AppendChar('x');
-            str.MakeReadOnly();
-            AuthenticationResult result = await app.AcquireTokenByUsernamePasswordAsync(TestConstants.Scope, TestConstants.User.Username, str).ConfigureAwait(false);
-
-            Assert.IsNotNull(result);
-            Assert.AreEqual("some-access-token", result.AccessToken);
-            Assert.IsNotNull(result.Account);
-            Assert.AreEqual(TestConstants.User.Username, result.Account.Username);
-
-            Assert.IsTrue(HttpMessageHandlerFactory.IsMocksQueueEmpty, "All mocks should have been consumed");
-        }
-
-        [TestMethod]
-        [TestCategory("PublicClientApplicationTests")]
-        public async Task ManagedUsernamePasswordAcquireTokenTest()
-        {
-            //add mock response for tenant endpoint discovery
-            HttpMessageHandlerFactory.AddMockHandler(new MockHttpMessageHandler
-            {
-                Method = HttpMethod.Get,
-                ResponseMessage = MockHelpers.CreateOpenIdConfigurationResponse(TestConstants.AuthorityOrganizationsTenant)
-            });
-
-            // user realm discovery
-            HttpMessageHandlerFactory.AddMockHandler(new MockHttpMessageHandler
-            {
-                Method = HttpMethod.Get,
-                ResponseMessage = new HttpResponseMessage(HttpStatusCode.OK)
-                {
-                    Content =
-                        new StringContent("{\"ver\":\"1.0\",\"account_type\":\"Managed\",\"domain_name\":\"id.com\"}")
-                },
-                QueryParams = new Dictionary<string, string>()
-                {
-                    {"api-version", "1.0"}
-                }
-            });
-
-            HttpMessageHandlerFactory.AddMockHandler(new MockHttpMessageHandler
-            {
-                Method = HttpMethod.Post,
-                ResponseMessage = MockHelpers.CreateSuccessTokenResponseMessage(),
-                PostData = new Dictionary<string, string>()
-                {
-                    {"grant_type", "password"},
-                    {"username", TestConstants.User.Username},
-                    {"password", TestConstants.DefaultPassword}
-                }
-            });
-
-            PublicClientApplication app = new PublicClientApplication(TestConstants.ClientId);
-
-            TokenCache cache = new TokenCache();
-
-            AuthenticationResult result = await app.AcquireTokenByUsernamePasswordAsync(TestConstants.Scope, TestConstants.User.Username, TestConstants.DefaultPassword).ConfigureAwait(false);
-
-            Assert.IsNotNull(result);
-            Assert.AreEqual("some-access-token", result.AccessToken);
-            Assert.IsNotNull(result.Account);
-            Assert.AreEqual(TestConstants.User.Username, result.Account.Username);
-
-            Assert.IsTrue(HttpMessageHandlerFactory.IsMocksQueueEmpty, "All mocks should have been consumed");
-        }
-
-        [TestMethod]
-        [TestCategory("PublicClientApplicationTests")]
-        public async Task ManagedUsernameSecureStringPasswordAcquireTokenTest()
-        {
-            SecureString str = new SecureString();
-            str.AppendChar('x');
-            str.MakeReadOnly();
-            //add mock response for tenant endpoint discovery
-            HttpMessageHandlerFactory.AddMockHandler(new MockHttpMessageHandler
-            {
-                Method = HttpMethod.Get,
-                ResponseMessage = MockHelpers.CreateOpenIdConfigurationResponse(TestConstants.AuthorityOrganizationsTenant)
-            });
-
-            // user realm discovery
-            HttpMessageHandlerFactory.AddMockHandler(new MockHttpMessageHandler
-            {
-                Method = HttpMethod.Get,
-                ResponseMessage = new HttpResponseMessage(HttpStatusCode.OK)
-                {
-                    Content =
-                        new StringContent("{\"ver\":\"1.0\",\"account_type\":\"Managed\",\"domain_name\":\"id.com\"}")
-                },
-                QueryParams = new Dictionary<string, string>()
-                {
-                    {"api-version", "1.0"}
-                }
-            });
-
-            HttpMessageHandlerFactory.AddMockHandler(new MockHttpMessageHandler
-            {
-                Method = HttpMethod.Post,
-                ResponseMessage = MockHelpers.CreateSuccessTokenResponseMessage(),
-                PostDataObject = new Dictionary<string, object>()
-                {
-                    {"grant_type", "password"},
-                    {"username", TestConstants.User.Username},
-                    {"password", TestConstants.DefaultPassword}
-                }
-            });
-
-            PublicClientApplication app = new PublicClientApplication(TestConstants.ClientId);
-
-            TokenCache cache = new TokenCache();
-
-           
-            AuthenticationResult result = await app.AcquireTokenByUsernamePasswordAsync(TestConstants.Scope, TestConstants.User.Username, str).ConfigureAwait(false);
-
-            Assert.IsNotNull(result);
-            Assert.AreEqual("some-access-token", result.AccessToken);
-            Assert.IsNotNull(result.Account);
-            Assert.AreEqual(TestConstants.User.Username, result.Account.Username);
 
             Assert.IsTrue(HttpMessageHandlerFactory.IsMocksQueueEmpty, "All mocks should have been consumed");
         }
